cache on NoteService methods ; clean code

+++ src/main/java/com/jeanchampemont/notedown/note/NoteService.java
@@ -22,7 +22,11 @@ import com.jeanchampemont.notedown.note.persistence.repository.NoteRepository;
 import com.jeanchampemont.notedown.security.AuthenticationService;
 import com.jeanchampemont.notedown.user.UserService;
 import com.jeanchampemont.notedown.user.persistence.User;
+import com.jeanchampemont.notedown.utils.exception.OperationNotAllowedException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.cache.annotation.CacheEvict;
+import org.springframework.cache.annotation.Cacheable;
+import org.springframework.cache.annotation.Caching;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -46,42 +50,87 @@ public class NoteService {
         this.authenticationService = authenticationService;
     }
 
+    /**
+     * @return all notes for the current user
+     */
     @Transactional(readOnly = true)
-    public Iterable<Note> getAll() {
+    public Iterable<Note> getNotes() {
         User user = authenticationService.getCurrentUser();
-        return getAll(user);
-    }
-
-    @Transactional(readOnly = true)
-    public Iterable<Note> getAll(User user) {
         return repo.findByUserOrderByLastModificationDesc(user);
     }
 
+    /**
+     * Get a not by ID.
+     * @param id
+     * @return
+     */
     @Transactional(readOnly = true)
+    @Cacheable("note")
     public Note get(UUID id) {
-        return repo.findOne(id);
-    }
-
-    @Transactional
-    public Note save(Note note) {
         User user = authenticationService.getCurrentUser();
-        return save(user, note);
+        Note note = repo.findOne(id);
+        if(note == null) {
+            return null;
+        }
+        if(hasReadAccess(user, note)) {
+            return note;
+        } else {
+            throw new OperationNotAllowedException();
+        }
     }
 
+    /**
+     * Create or update a note
+     * @param note
+     * @return
+     */
     @Transactional
-    public Note save(User user, Note note) {
-        note = updateLastModification(note);
-        note.setUser(user);
-        return repo.save(note);
+    @CacheEvict(value = "note", key = "#note.id")
+    public Note createUpdate(Note note) {
+        User user = authenticationService.getCurrentUser();
+        Note originalNote = repo.findOne(note.getId());
+        if(originalNote == null) {
+            originalNote = note;
+        }
+        if(hasWriteAccess(user, originalNote)) {
+            originalNote = updateLastModification(originalNote);
+            originalNote.setTitle(note.getTitle());
+            originalNote.setContent(note.getContent());
+            originalNote.setUser(user);
+            return repo.save(originalNote);
+        } else {
+            throw new OperationNotAllowedException();
+        }
     }
 
+    /**
+     * Delete a note by id
+     * @param id
+     */
     @Transactional
+    @CacheEvict(value = "note")
     public void delete(UUID id) {
-        repo.delete(id);
+        User user = authenticationService.getCurrentUser();
+        Note originalNote = repo.findOne(id);
+        if(originalNote != null) {
+            if (hasWriteAccess(user, originalNote)) {
+                repo.delete(id);
+            } else {
+                throw new OperationNotAllowedException();
+            }
+        }
     }
 
     private Note updateLastModification(Note note) {
         note.setLastModification(new Date());
         return note;
+    }
+
+    private boolean hasReadAccess(User user, Note note) {
+        return note.getUser() == null || note.getUser().getId() == user.getId();
+    }
+
+    private boolean hasWriteAccess(User user, Note note) {
+        return note.getUser() == null || note.getUser().getId() == user.getId();
     }
 }
